+++ client/src/gui/toolkit/SDL_gfx.c
@@ -4260,7 +4260,7 @@ int aaellipseColor(SDL_Surface * dst, Sint16 x, Sint16 y, Sint16 rx, Sint16 ry, 
 
         /* Calculate alpha */
         if (s != 0.0) {
-            cp = abs(d) / abs(s);
+            cp = (double) abs(d) / (double) abs(s);
             if (cp > 1.0) {
                 cp = 1.0;
             }
@@ -4319,7 +4319,7 @@ int aaellipseColor(SDL_Surface * dst, Sint16 x, Sint16 y, Sint16 rx, Sint16 ry, 
 
         /* Calculate alpha */
         if (t != 0.0) {
-            cp = abs(d) / abs(t);
+            cp = (double) abs(d) / (double) abs(t);
             if (cp > 1.0) {
                 cp = 1.0;
             }
